Enable custom logic to determine worker configuration
Previously, configuration could only be loaded from a
YAML file, or a Hash instance.

This change introduces the concept of a *configuration loader*.

All of the previous configuration logic has been moved to the
FileOrHashLoader class, which will be used by default.

To provide your own logic, you set the `configuration_loader`
class variable on Resque::Pool to an object that
responds to `#call` (which can simply be a lambda/Proc).

See the README for more details.