Support no-pidfile mode
This change makes it possible to run multiple copies of resque-pool on
the same server using the --daemon flag, which was previously not
allowed since the server refuses to start if its configured pidfile
already exists.  Process supervisors like upstart and systemd already
know what pids they are managing, so in this context a pidfile is
unnecessary.

Without the ability to run concurrent daemon instances, zero-downtime
deployment (in which a new resque-pool instance starts up and slowly
replaces the previous instance until it is done) is not feasible.