Fix occasional all_resque_pool_processes deadlock
`Open3.capture3` is vulnerable to deadlock because it has two threads
doing blocking reads on separate stderr/stdout pipes, each with limited
buffer size.

In this case, we only need stdout to parse the pids. We want stderr in
case the ps command fails. So we get the best of both worlds by
redirecting stderr to stdout, relying on exitstatus to decide whether
the command succeeded, and showing full output if it didn't.