🔒 Configure Trusted Publisher
This was copied from what I've been using with `net-imap`.

The action will automatically trigger when `v*` tags are pushed.  And
I've configured the `rubygems.org` environment to require a reviewer to
confirm, before actually executing the action (self-review _is_
allowed).  It attempts to creates a draft release after the deploy
completes.

GitHub wouldn't let me add the `resque` "Core" team as reviewers... but
I've also configured the environment so that organization admins can
bypass the protection rules, so I suppose that doesn't really matter!

This way, in the extremely unlikely event that an emergency release is
required and I can't be contacted (or I've moved on, without a proper
handoff), any `resque` organization admin can make a new release by
simply bumping the `VERSION` contant, pushing the tag, and reviewing
the action.