🐛 Fix `Logging.reopen_logs!` for ruby 3.4
I think this is probably a bug in ruby 3.4's `IO#reopen`, but it's not a
big deal to explicitly copy over the original IO encodings.

It should be safe to delete that old code that handled IO differently
from stdin, stdout, stderr: the referenced ruby bug was fixed 11 years
ago and released with ruby 2.0.0.